requirements.txt
numpy>=1.26.0
pandas>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
influxdb-client>=1.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
